Refactor: rename backend source files for clarity
exec_core.rs  → interp.rs    (interpreter path)
jit_a64.rs    → jit_arm64.rs (AArch64 JIT)
jit_x64.rs    → jit_amd64.rs (x86-64 JIT)

Fixes #8
